requirements.txt
torchvision==0.6.0
torch==1.5.0
Flask==1.1.2
Pillow==8.2.0
